Conflicts:
	nbproject/private/private.properties
	nbproject/private/private.xml
	src/conf/persistence.xml
	src/java/br/ufjf/controller/Controller.java
	src/java/br/ufjf/dao/ClienteDAO.java
	src/java/br/ufjf/dao/ExamesDAO.java
	src/java/br/ufjf/dao/TipoExameDAO.java
	src/java/br/ufjf/dao/UsuarioDAO.java
	src/java/br/ufjf/model/Cliente.java
	src/java/br/ufjf/model/Endereco.java
	src/java/br/ufjf/model/TipoExame.java
	src/java/br/ufjf/model/Usuario.java
	web/css/main.css
	web/pages/administracao.jsp
	web/pages/cadastroUsuarios.jsp
	web/pages/resultados.jsp

+++ src/java/br/ufjf/dao/UsuarioDAO.java
@@ -11,13 +11,17 @@ import javax.persistence.Query;
 
 public class UsuarioDAO {
 
+    String comparalogin;
+    
     public EntityManager getEM() {
         EntityManagerFactory factory = Persistence.createEntityManagerFactory("LaboratorioPU");
         return factory.createEntityManager();
     }
 
-    //Create e insert junto....
+    //Create e alteração junto....
     public Usuario salvar(Usuario usuario) {
+        
+        UsuarioDAO dao = new UsuarioDAO();
 
         EntityManager em = getEM();
 
@@ -25,10 +29,10 @@ public class UsuarioDAO {
 
         if (usuario.getLogin() == null) {//verifica existe...
 
-            em.persist(usuario); // executa insert
+            em.persist(usuario); // executa insert se não existir
 
         } else {
-            usuario = em.merge(usuario); //executa update
+            usuario = em.merge(usuario); //executa update se existir e faz o merge...
 
         }
         em.getTransaction().commit();
@@ -36,27 +40,35 @@ public class UsuarioDAO {
 
         return usuario;
     }
-    
+
     //Delete Usuario... 
-    public void removerUsuario(String login) {
+    public void removerUsuario(Usuario matricula) {
 
         EntityManager em = getEM();
-        Usuario usuario = em.find(Usuario.class, login);
         
+        Usuario usuario = em.find(Usuario.class, matricula);
+
         try {
             em.getTransaction().begin();
-            em.remove(usuario);
+
+            if (usuario.getMatricula().equals(matricula)) {
+                em.remove(usuario);
+            }
+            
             em.getTransaction().commit();
+            
         } finally {
             em.close();
         }
     }
+
     //Procura usuario por login...
+
     public Usuario consultaUsuarioporlogin(String login) {
 
         EntityManager em = getEM();
         Usuario u = null;
-        
+
         try {
             u = em.find(Usuario.class, login); // executa select
         } finally {
@@ -64,7 +76,6 @@ public class UsuarioDAO {
         }
         return u;
     }
-
 
     public Usuario getUsuarioByLoginAndPassword(String login, String senha) {
         Query q = getEM()
